sending logs to otel
