+++ backend/src/zango/core/monitoring/tasks.py
@@ -1,9 +1,5 @@
 from celery.signals import worker_process_init
 
-from .telemetry import setup_telemetry
-from .logging import setup_logging
-
-
 @worker_process_init.connect
 def initialize_otel(**kwargs):
     """
@@ -13,6 +9,8 @@ def initialize_otel(**kwargs):
     when a worker process starts. It sets up the telemetry and logging 
     configurations.
     """
+    from .telemetry import setup_telemetry
+    from .logging import setup_logging
     setup_telemetry(
         add_django_instrumentation=False
         )
