+++ backend/src/zango/core/monitoring/utils.py
@@ -1,4 +1,6 @@
 import os
+import logging
+from opentelemetry.sdk._logs import LoggingHandler
 
 
 def otel_is_enabled():
@@ -24,3 +26,46 @@ def otel_otlp_protocol():
 
 def otel_resource_name():
     return str(os.getenv("OTEL_RESOURCE_NAME", "Zango"))    
+
+def _get_tenant_name():
+    from django.db import connection
+    try:
+        tenant_name = connection.tenant.name
+    except Exception as e:        
+        tenant_name = 'FakeTenant'
+    return tenant_name
+
+def _get_tenant_filename(original_name):
+    if original_name.startswith("pluginbase"):
+        result = f"workspaces.{_get_tenant_name()}.{'.'.join(original_name.split('.')[3:])}"
+    else:
+        result = original_name        
+    return result
+           
+def get_loguru_format(record):
+    file_name = _get_tenant_filename(record['name'])
+    
+    loguru_format = (
+        f"<magenta>{os.getpid()}|</magenta>"
+        "<green>{time:YYYY-MM-DD HH:mm:ss.SSS}</green>|"
+        "<level>{level}</level>|"
+        f"<cyan>{file_name}</cyan>:"
+        "<cyan>{function}</cyan>:"
+        "<cyan>{line}</cyan> - "
+        "<level>{message}</level>\n"
+    )
+    return loguru_format
+
+class LogGuruCompatibleLoggerHandler(LoggingHandler):
+    
+    def emit(self, record: logging.LogRecord) -> None:
+        # The Otel exporter does not handle nested dictionaries. Loguru stores all of
+        # the extra log context developers can add on the extra dict. Here unnest
+        # them as attributes on the record itself so otel can export them properly.
+        for k, v in record.extra.items():
+            setattr(record, f"zango.{k}", v)
+        del record.extra
+
+        # by default otel doesn't send funcName, rename it so it does.
+        setattr(record, "python_function", record.funcName)
+        super().emit(record)
